♻️ refactor: move to `utils/client`
