+++ src/services/chat.ts
@@ -41,7 +41,7 @@ import { createErrorResponse } from '@/utils/errorResponse';
 import { FetchSSEOptions, fetchSSE, getMessageError } from '@/utils/fetch';
 import { genToolCallingName } from '@/utils/toolCall';
 import { createTraceHeader, getTraceId } from '@/utils/trace';
-import { parsePlaceholderVariablesMessages } from '@/utils/parserPlaceholder';
+import { parsePlaceholderVariablesMessages } from '@/utils/client/parserPlaceholder';
 
 import { createHeaderWithAuth, createPayloadWithKeyVaults } from './_auth';
 import { API_ENDPOINTS } from './_url';
