+++ src/store/user/slices/auth/selectors.ts
@@ -23,20 +23,20 @@ const nickName = (s: UserStore) => {
 };
 
 const username = (s: UserStore) => {
-  const defaultUserName = s.user?.username || s.user?.fullName;
   if (!enableAuth) {
     if (isDesktop) {
-      return defaultUserName;
+      return s.user?.username;
     }
     return DEFAULT_USERNAME;
   }
 
-  if (s.isSignedIn) return defaultUserName;
+  if (s.isSignedIn) return s.user?.username;
 
   return 'anonymous';
 };
 
 export const userProfileSelectors = {
+  fullName: (s: UserStore): string => s.user?.fullName || '',
   nickName,
   userAvatar: (s: UserStore): string => s.user?.avatar || '',
   userId: (s: UserStore) => s.user?.id,
